Migrate to AndroidX, removed ButterKnife, restructure libraries
