+++ library-common/build.gradle
@@ -5,5 +5,5 @@ android {
 }
 
 dependencies {
-    api 'com.android.support:appcompat-v7:27.0.2'
+    api 'androidx.appcompat:appcompat:1.6.1'
 }
