+++ library-common/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="cz.ursiny.countertextview"/>
+          package="cz.ursiny.countertextview.library.common"/>
