+++ sample-rxjava1/src/main/java/cz/ursiny/countertextview/sample1/MainActivity.java
@@ -1,41 +1,34 @@
 package cz.ursiny.countertextview.sample1;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.widget.SeekBar;
-import android.widget.TextView;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.view.View;
 
 import com.jakewharton.rxbinding.widget.RxSeekBar;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import cz.ursiny.countertextview.library.CounterTextView;
+import cz.ursiny.countertextview.sample1.databinding.ActivityMainBinding;
 import rx.Subscription;
 
 public class MainActivity extends AppCompatActivity {
-
-    @BindView(R.id.counter)
-    CounterTextView mCounter;
-    @BindView(R.id.seekBar)
-    SeekBar mSeekBar;
-    @BindView(R.id.target)
-    TextView mTarget;
     private Subscription mSubscription;
+    private ActivityMainBinding binding;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        ButterKnife.bind(this);
+        binding = ActivityMainBinding.inflate(getLayoutInflater());
+        View view = binding.getRoot();
+        setContentView(view);
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-        mSubscription = RxSeekBar.userChanges(mSeekBar)
+        mSubscription = RxSeekBar.userChanges(binding.seekBar)
                 .map(Long::valueOf)
-                .doOnNext(aLong -> mTarget.setText(String.valueOf(aLong)))
-                .subscribe(mCounter.targetAction());
+                .doOnNext(aLong -> binding.target.setText(String.valueOf(aLong)))
+                .subscribe(binding.counter.targetAction());
     }
 
     @Override
